requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-cors-headers==4.3.1
google-api-python-client==2.181.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib==1.1.0
python-decouple==3.8